+++ back-strapi/api/content/controllers/content.js
@@ -1,6 +1,6 @@
-'use strict';
+"use strict";
 
-const { sanitizeEntity } = require('strapi-utils');
+const { sanitizeEntity } = require("strapi-utils");
 
 /**
  * Read the documentation (https://strapi.io/documentation/developer-docs/latest/development/backend-customization.html#core-controllers)
@@ -8,46 +8,67 @@ const { sanitizeEntity } = require('strapi-utils');
  */
 
 module.exports = {
-	async updateForMobileApp(ctx) {
-		const userFromContext = ctx.state.user; 
-		const body = ctx.request.body;
+  async updateForMobileApp(ctx) {
+    const userFromContext = ctx.state.user;
+    const body = ctx.request.body;
 
-		const content = await strapi.services['content'].findOne({ title: body.title });
-		if (!content) {
-			return ctx.badRequest(null, 'Content [' + body.title + '] does not exist');
-		}
+    const content = await strapi.services["content"].findOne({
+      title: body.title,
+    });
+    if (!content) {
+      return ctx.badRequest(
+        null,
+        "Content [" + body.title + "] does not exist"
+      );
+    }
 
-		let thematiqueMobile = await strapi.services['thematique-mobile'].findOne({title: body.theme_app})
-		if (!thematiqueMobile) {
-			thematiqueMobile = await strapi.services['thematique-mobile'].create({
-				title: body.theme_app
-			})
-		}
+    let thematiqueMobile = await strapi.services["thematique-mobile"].findOne({
+      title: body.theme_app,
+    });
+    if (!thematiqueMobile) {
+      thematiqueMobile = await strapi.services["thematique-mobile"].create({
+        title: body.theme_app,
+      });
+    }
 
-		let level = {}
-		if (body.level) {
-			level = await strapi.services['niveau'].findOne({value: parseInt(body.level)})
-			if (!level) {
-				level = await strapi.services['niveau'].create({
-					value: parseInt(body.level),
-					name: 'Niveau ' + body.level
-				})
-			}
-		}
+    let etiquette = await strapi.services["etiquette"].findOne({
+      title: body.etiquette,
+    });
+    if (!etiquette) {
+      etiquette = await strapi.services["etiquette"].create({
+        title: body.etiquette,
+      });
+    }
 
+    let level = {};
+    if (body.level) {
+      level = await strapi.services["niveau"].findOne({
+        value: parseInt(body.level),
+      });
+      if (!level) {
+        level = await strapi.services["niveau"].create({
+          value: parseInt(body.level),
+          name: "Niveau " + body.level,
+        });
+      }
+    }
 
-		let updatedContent = {
-			title_mobile: content.title,
-			thematique_mobile: thematiqueMobile.id,
-			niveau: level.id
-		}
+    let updatedContent = {
+      title_mobile: content.title,
+      thematique_mobile: thematiqueMobile.id,
+      etiquette: etiquette.id,
+      niveau: level.id,
+    };
 
-		if (body.new_title) {
-			updatedContent.title_mobile = body.new_title
-		}
+    if (body.new_title) {
+      updatedContent.title_mobile = body.new_title;
+    }
 
-		const entity = await strapi.services['content'].update({id: content.id}, updatedContent)
+    const entity = await strapi.services["content"].update(
+      { id: content.id },
+      updatedContent
+    );
 
-		return sanitizeEntity(entity, { model: strapi.models.content });
-	}
+    return sanitizeEntity(entity, { model: strapi.models.content });
+  },
 };