+++ tumeplayMobile/src/services/api/contact.js
@@ -1,19 +1,19 @@
-const REACT_APP_API_URL = process.env.REACT_APP_API_URL;
+import {REACT_APP_URL} from '@env';
 
 const ContactsAPI = {
   postContact: async userAdress => {
-    await fetch(REACT_APP_API_URL + '/contacts', {
+    await fetch(REACT_APP_URL + '/contacts', {
       method: 'POST',
       headers: {
         Accept: 'application/json',
         'Content-Type': 'application/json',
       },
       body: JSON.stringify({
-        name: userAdress.firstName,
-        email: userAdress.emailAdress,
+        name: userAdress.first_name,
+        email: userAdress.email,
         zipcode: userAdress.zipCode,
         box: userAdress.box_id,
-        type: userAdress.type
+        type: userAdress.type,
       }),
     });
   },