+++ tumeplayMobile/src/views/Box.js
@@ -50,8 +50,7 @@ const Box = ({navigation}) => {
         <View style={styles.subtitleContainer}>
           <Text style={styles.congrats}>Bravo !</Text>
           <Text style={styles.description}>
-            Tu as assez de points pour commander un kit de ton choix (CHANGER LE
-            TEXT)
+            Tu peux maintenant commander un kit
           </Text>
         </View>
       </View>